refactor(db): split web.ts into dexie-database.ts and service index
The 910-line web.ts contained two distinct concerns: historical Dexie
schema migrations (V1-V6) that will never change, and the active CRUD
service implementation. Splitting isolates the dead migration code so
it can't be accidentally touched while editing the service.

  services/db/web.ts → services/db/web/
  ├── dexie-database.ts  (WhisperingDatabase class + V1-V6 migrations)
  └── index.ts           (createDbServiceWeb + audio helpers)

Zero import changes—`./web` resolves to `./web/index.ts` identically.
